chore: refactor Condition
