+++ src/depot_server/db2/models/item/item.py
@@ -4,8 +4,8 @@ from enum import StrEnum
 from tortoise import fields
 from tortoise.models import Model
 
-from depot_server.db2.common import Condition
 from depot_server.db2.models.asset.asset import Asset
+from depot_server.db2.models.common import Condition
 from depot_server.db2.models.item.item_group import ItemGroup
 from depot_server.db2.models.item.storage_location import StorageLocation
 from depot_server.db2.models.report.report_profile import ReportProfile
