+++ src/depot_server/db2/models/report/report_element.py
@@ -1,7 +1,7 @@
 from tortoise import fields
 from tortoise.models import Model
 
-from depot_server.db2.common import UserIdField
+from depot_server.db2.models.common import UserIdField
 
 
 class ReportElement(Model):
